backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
openai
requests
chromadb
sentence-transformers
pypdf